requirements.txt
flask
requests
beautifulsoup4
faiss-cpu
transformers
sentence-transformers
torch
